devicepkg-dev: ship compiler-gcc.h for GCC8
We used to put a "compiler-gcc6.h" file inside every vendor kernel's
aport. This is redundant and only works with GCC6. With this patch,
a common "compiler-gcc.h" file will be used, that works independently
of the GCC version.

The file is from Decatf's android_kernel_samsung_p4 repository, where
it was taken from upstream Linux, but adjusted to work with older
vendor kernels (different #ifndef check at the top etc).

https://github.com/Decatf/android_kernel_samsung_p4/blob/bd0772df2132bf56cda78e7204facf434cd5aa40/include/linux/compiler-gcc.h