motorola-def: add basic support for Moto One Hyper
this adds basic functionality for the Moto One Hyper.
the device boots with rootfs on sd, booted by 'pmbootstrap flasher boot'.
usbnet works, and i can ssh in. there is no framebuffer device, but
weston works with the drm backend instead of fbdev. touchscreen support
is not implemented yet, but i will update this with support when i get
the chance. most other functionality either does not seem to work yet or
has not been tested (wifi, audio, modem, etc).