temp/gtk+3.0: fork from Alpine to apply Purism mobile patches (MR 1312)
Purism has some patches to make GTK apps behave better on mobile screens
and dimensions. These break desktop usage so will not be applied to the
edge branch for now, but we should have these when shipping pmOS on the
PinePhone with the stable branch.

Please test applications known to be sized wrongly before, they should
look a lot better now.

The build hangs if -j > 1, so this package disables parallel make jobs when
building on these architectures. Everything else uses the default # of
parallel jobs for the CPU/system.