main/linux-postmarketos-allwinner: *** sync with master ***
Cherry-picking has gone wrong before, it was done in the wrong order,
and while commit 1e9d04 ("enable xt_owner and xt_time kernel modules (MR 1343)")
was picked, the changes in the kernel config from that commit are
missing on current v20.05 HEAD.

This patch copies the current kernel config from master to the v20.05
branch and thereby makes the v20.05 version equal with master. The only
difference is now the pkgrel, but this will also go away with the next
upgrade in the kernel version.

When cherry-picking patches, we should always cherry-pick all patches
from linux-postmarketos-allwinner from master to v20.05 and in the right
order. Then we don't have complicated diffs to resolve and errors like
this one happening, and we can be sure that the version on v20.05 is the
same that was tested on master.