temp: add or1k-elf cross-embedded toolchain
Fork gcc-cross-embedded and dependencies from Alpine edge. Limit it to
only build or1k-elf from x86_64, because that's what we need to build the
crust firmware for the pinephone. More architectures could be enabled on
demand. I've verified that this builds before pushing, and that it's
possible to build crust from the open merge request with this toolchain.

(cherry picked from commit 59a374f6f93304647a8851c24574ad463ca7830b)
Note: picked from v20.05 branch, and not tested individually. Not needed
since the initial build of the branch for arm isn't done yet. Removed
binutils-cross-embedded, because that's available in Alpine v3.13.