temp/newlib-stage1: fix depends
Make the newlib-stage1 meta-package depend on newlib-or1k-elf-stage1
instead of newlib-or1k-elf. Otherwise, the meta-package could not be
installed in postmarketOS, v20.05 branch, because we have only forked
newlib-stage1 and not newlib. This is a bug in the upstream aport too.

Related: https://gitlab.alpinelinux.org/alpine/aports/-/merge_requests/10318