main/postmarketos-ondev: upgrade to 0.3.2 (MR 1780)
The calamares module and branding was made distro-agnostic and moved to
calamares-extensions.git. A new "make install" target inside
postmarketos-ondev.git installs all files, so we need less logic in the
APKBUILD. Lightdm is replaced with tinydm.

Bring back the dependency on plasma-phone-components-vkbd-style, so we
have a mobile keyboard instead of the default desktop keyboard.

Add xset dependency to disable screen blanking.