temp/megapixels: upgrade to 0.9.0 (MR 1612)
Set arch to all, as it builds for more than just ARM and the author
(Martijn) doesn't plan to make it ARM specific forever. We'd like to
ship this as default in postmarketos-ui-phosh, which is easier with
arch=all as well.

Related: https://gitlab.alpinelinux.org/alpine/aports/-/merge_requests/13006