main/postmarketos-base: post-install: unroll rc-update loops (MR 1807)
Prepare for proper linting and dependency test of -openrc subpackages.

Replace setup-udev call with the 'rc-update add' calls it is effectively
performing in the /etc/runlevels loop. This also stops the
postmarketos-base.post-install script from trying to start udev, which
isn't useful as that would always get executed while the installation is
being prepared in pmbootstrap.