device/pine64-pinephone: Install default alsa conf to /etc/alsa/conf.d/ file
Instead of installing the alsa conf to /etc/asound.conf this changes the
install path to be /etc/alsa/conf.d/00_alsa_pinephone_dshare.conf. This is
preferable because if you install to /etc/asound.conf as was done previously,
any (other) user or package configuration in /etc/alsa/conf.d/ is ignored.

This change allows other packages to add additional alsa configurations in
/etc/alsa/conf.d as well, restoring Sxmo's dmix/dsnoop setup.