device-pine64-pinephone: add elogind suspend hook for managing wifi pm
This adds an elogind hook for enabling PM right before suspend, then
disabling it on resume to allow for faster wifi reconnect. When the
device is operating on battery power, pm is re-enabled after 30 seconds.

In my testing, 5 seconds was not enough time for it to reconnect to my
AP, and 30 seconds seemed like the right amount. Having PM disabled for
a few extra seconds seems like a reasonable price to pay for having a
faster reconnect without having to worry about PM being switched on
again half way through the reconnection.

I didn't see any obvious problems with having a `sleep` in the hook,
elogind seems to fork this and it didn't appear to block anything. If
more hooks are added later, it might block execution of future hooks,
but let's address that if/when the need arises.

fixes #973