temp/modemmanager: upgrade to 1.16.2 (use upstream rpmsg patch)
When upstreaming the RPMSG patch for modem functionality on devices
with MSM8916/MSM8974 SoC, it was decided that it is cleaner to let
the bam-dmux driver operate in Raw-IP mode rather than Ethernet mode
(which is partially broken). There were also several other changes made
that require minor changes in the bam-dmux driver.

Because of that, the upstreamed patch now depends on some changes
in the bam-dmux driver (and vice-versa) so we need to upgrade
ModemManager and the kernels together.

The commit updates the ModemManager package in pmaports to the latest
release.