temp/u-boot-librem5: use pmb:cross-native (MR 1988)
Let pmbootstrap build the package in the native chroot. This is
important for the v21.03 branch, which needs to ship its own
cross-embedded toolchain, as the one from Alpine is only in testing and
therefore not available in the Alpine v3.13 branch that pmOS v21.03 is
based on. By using the native branch, it is enough to build the
cross-embedded toolchain for x86_64, and we can avoid qemu related
errors when attempting to build it for arm architectures.

As a nice benefit, this makes compilation of the package faster on edge,
too.

[ci:skip-build]: already built successfully in CI

(cherry picked from commit fcc2ef8f8602b288f23dd5545c7de984be8c8df8)