device-purism-librem5: Fix ship mode init script
The script from purism expects to run under systemd, so it bailed early
when called without a 'poweroff' parameter.

This moves the ship mode script from purism into the openrc init script,
and fixes some issues to make it run in a posix shell.

It's not expected that this script will by changed much by purism moving
forward, so moving it into this seems 'safe'

i2c-tools is required for the ship mode script (which calls i2cset)