kde/kpeoplesink: disable armv7, remove from _pmb_recommends
Alpine v3.13 doesn't have the sink and flatbuffers for armv7, so it
fails to build. Disable to finish up the initial build of v21.03 for
armv7. Remove it from _pmb_recommends of postmarketos-ui-plasma-mobile,
so plamo can still be installed for armv7.

Fixing the depends and reverting this is discussed in #996.