temp/geary: upgrade to 3.38.2+git20210118+7bb025-0pureos3 (MR 2035) (MR 2130)
Latest patches from Purism, including some they backported from
upstream.

fixes #1058

The debian dir was dropped from the archives in this fork repo, but the
number of patches has been dramatically reduced.

A new mechanism (thanks @ollieparanoid) is added here to compare the
list of patches stored in pmaports when this package was upgraded to the
list of patches in purism's repo @ some commit this package uses, so
that any difference can be detected (pacakge will fail to build). If
there's a mismatch, the patches from this repo should be updated/added
to match what's in purism's repo @ the commit, and purism_patches.series
should be updated to match the debian/patches/series file in the purism
repo.

(cherry picked from commit ffa73f09af288504595b85651e9229c1171c8a92)