temp/u-boot-pinephone: lower DRAM clock to 528
Some PinePhones need the DRAM clock frequency lowered to 528. This
simple patch for the v21.03 branch just hardcodes the lower value. A
more sophisticated patch is being developed in MR 1998 (will allow users
to select the frequency in the u-boot update script, and add the
currently booted frequency as kernel boot parameter).