main/postmarketos-mkinitfs: add support for keyboard detect script (MR 2327)
This uses a script, /usr/bin/phys-keyboard-present, that is added to the
initfs for determining whether or not to run osk-sdl with -x to hide the
on-screen keyboard.

phys-keyboard-present is an (optional) script provided by device
packages, since the implementation for detecting whether or not a
keyboard is present will likely differ across different devices. The
script exit with a return code of 0 if a physical keyboard is present,
or non-zero if one is *not* present.

This patch requires osk-sdl >=0.65

Note: shellcheck set to ignore SC2086 because we actually want to split
on spaces here since each item is a separate parameter to osk-sdl.
See: https://github.com/koalaman/shellcheck/wiki/SC2086#exceptions