main/postmarketos-base: depend on wpa_supplicant
Peope are actually using our -none UI which is awesome. However, using
that it's impossible to setup the system when you don't have ethernet
access nearby as there are no tools installed to configure networking.
I've asked the Alpine devs and on the Alpine minimal image
wpa_supplicant is pre-installed for this, and I think we should do the
same. It doesn't hurt installing it for UI's (most of the time it gets
installed there through other means anyway) and would really help the
-none UI.

Preferably this goes into a _pmb_recommends block but that is not
supported in non-ui packages.