main/postmarketos-ui-sxmo-*: depend on pmos-base-ui package
Without this explicit dependency, installing sxmo with the 'qemu' device
fails:

    # apk add device-qemu-amd64 postmarketos-ui-sxmo-de-sway
    ERROR: unable to select packages:
    Huh? Error reporter did not find the broken constraints.

I don't understand *why* it fails, but something in the dependency
resolution or ?? for these two packages breaks down. The fact that the
pmos-ui-sxmo-* packages depend on sxmo-common, which depends on
pmos-base-ui doesn't seem to make any difference.

This failure only seems to be specific to device-qemu-*, maybe because
that device package has an explicit dependency on postmarketos-base?

Fixes #1343