main/postmarketos-ui-xfce4: add package "procps"
The package "procps" contains a more feature-rich command "ps" than
the one from busybox. This allows xfce4-session to read the commands
of running programs, which is needed to restore them at next login
(if the default option "Save session for future logins" is set at the
shutdown/reboot dialog).