pine64-pinephonepro: switch to extlinux booting only
This removes u-boot from our image and allows booting via an
extlinux-compatible platform firmware instead. And hey, we already ship
that anyway! So to flash pmOS on the PPP now (sd-card only for now) we
will require the user to flash a platform firmware image once after
which they can write this pmOS image to an sdcard.

I recommend using https://github.com/Tow-Boot/Tow-Boot as the platform
firmware of choice. Once flashed to internal storage, put pmOS on an
sd-card and put it in the device, then boot the device while holding the
volume-down button

Support for installing to eMMC support will be done in installers like
pmbootstrap and the on-device installer