main/postmarketos-ui-framebufferphone: remove
This depends on too many packages that are still in Alpine testing, so
we decided to remove it. If somebody wants to bring it back to stable,
create a MR with all missing dependencies to v22.06, and we can include
it in a service pack.

Related: https://gitlab.alpinelinux.org/alpine/aports/-/issues/13857