nokia-n900: drop weston subpackage (MR 3174)
Drop the weston config, as it was for fbdev which has not been supported
in weston anymore for a while.

Remove Martijn as Co-Maintainer, as both previous maintainers of the
Nokia N900 stopped maintaining it when the device was briefly dropped
from community.

Co-Authored-By: Oliver Smith <ollieparanoid@postamrketos.org>
[ci:skip-build]: already built successfully in CI

(cherry picked from commit ba3731956dfb4aa8848dd576c5578009ddc33a2a)