main/postmarketos-welcome: version provides (MR 3208)
There was lately a discussion upstream due to apk 2.12.10 breaking
some packages in aports[1]. The conclusion in IRC is that provides
were designed to be either versioned or have a provider_priority.
Otherwise, the provides will be treated as a "virtual package".
Which is not what we want here.

[1] https://gitlab.alpinelinux.org/alpine/apk-tools/-/issues/10847

(cherry picked from commit bec9861ec4745b042e4011f45dbe1890d96b57d4)