main/device-pine64-pinephone: switch to UEFI booting
Drop U-Boot scripts and enable UEFI booting with grub on the
PinePhone to be EBBR compliant. This brings the PinePhone boot process
on par with the PinePhone Pro. Tow-Boot is now a requirement to even be
able to boot as U-Boot does not have UEFI enabled.

Dropped the symlink from .. to /boot/allwinner because symlinks are
not supported on fat32 which is required for booting into UEFI.