CI: aports version check: modernize with new rules (MR 3615)
Update the script once more with the goal that we no longer need to
comment on pkgrel/pkgver bumps. Parse not only the version of the old
and new package, but also the checksums and remote sources.

Replace the previous check of simply having $pkgver-r$pkgrel higher
than the upstream APKBUILD with more detailed rules:

  APKBUILD has remote sources?      ------>       Did sha512sums change?
   |                                  No            |         |
   | Yes                                            | No      | Yes
   v                                                |         v
  Did the file name of a remote                     |     * Increase pkgver
  source change?                    ----.           |     * Set pkgrel=0
   |                                    |           |
   | Yes                                | No        |
   v                                    v           v
  Is pkgver=9999?         ----->   * Don't change pkgver
  (Forked from Alpine)     Yes     * Bump pkgrel by 1
   |
   | No
   v
  * Set the pkgver to the version
    of the source tarball
  * Set pkgrel=0

I'll also put this in the wiki at:
https://postmarketos.org/howto-bump-pkgrel-pkgver

Besides the new rules, print a detailed error message below each package
where the test failed that explains exactly what was expected and what
the reasoning behind that is. Also add a hint that users can run this
locally with 'pmbootstrap ci' now.

Use logging instead of print, so we get colored ERROR: and NOTE:
statements from pmbootstrap code, this should make it easier to read.