main/pmos-ui-phosh: use superd for user services
This uses gnome-session to run superd by changing the directory that
gnome-session searches for apps to autostart to one that has only a
.desktop file that runs superd. superd then starts/manages non-critical
services.

Some services in /etc/xdg/autostart are *critical* to
phosh/gnome-session, they are listed in a phosh session file.
gnome-session will continue to use /etc/xdg/autostart when searching for
these services, and gnome-session is still used to start those. They are
"critical" because stopping any of them causes the session to die. As
such, I do not believe it makes sense to manage those in superd anyways,
since if any die, the whole session (including superd) is doing down
with it.