purism-librem5: disable power management for Sparklan/brcmfmac wifi (MR 3793) (MR 3900)
This wifi module doesn't ship with the L5, but I have one because it's
generally more stable than the redpine thing that does ship with the
L5... however there's a known problem with having PM enabled on it.

This adds a NM config option, that matches specifically on this
device/driver, to disable wifi. PM for the redpine should not be
affected by this config change.

[ci:skip-build] already built successfully in CI

(cherry picked from commit c3669975e92f58de071d9b4736af4a24f11bb89f)