pmos-ui-phosh: add mmsd-tng as a dependency
Even though many carriers don't use/offer MMS, having mmsd around is
useful since it will clear out any SMS WAPs that might be received. If
the queue is "too full", it can lead to no longer being able to receive
any SMS (some might recall seeing this problem from a couple of years
ago...). So I think it's a good idea to always run mmsd to prevent these
from building up. And, obviously, it's helpful to have it running for
those of us who do use MMS so that it requires a little less user
interaction to enable it.