postmarketos-mkinitfs-hook-kernelmodules: new aport
This hook is used to copy newer kernel modules that are included in the
ramdisk into the rootfs during boot. Effectively allowing for new
kernels that have been built in CI to be "installed" seamlessly on
Android devices by just flashing/booting a new boot image.

This is intended for use in CI or for local kernel development.