community/soc-qcom-sdm845: depend on msm-modem-ipa
This adds a new udev rule to disable waking up from suspend when the IPA
receives an interrupt. This may be the cause of some "random" wakeups
that people have reported on SDM845 devices. Until we get proper support
for handling these and filtering packets in the IPA itself, this avoids
those random wakeups and potential associated power drain.