main/postmarketos-initramfs: rework logging
Make logs always go to kmsg, so that if you have a console configured
then you'll always be able to see the init script output. The
behaviour of PMOS_NO_OUTPUT_REDIRECT is adjusted so that it just
decides whether or not openrc will log to the console or not.

The original stdout/stderr fds are preserved and later restored, this is
important for debug-shell to run on the TTY.