u-boot-veyron: new aport
This packages u-boot.rom files for 4 boards supported in U-Boot
(mickey, jerry, minnie, speedy). This file is about to be flashed
to SPI flash and replace stock bootloader.

I tested it only on speedy, so I promote this only on speedy wiki
page.

I've tried to package it so that if a new board is added to U-Boot,
we only have to add another value to $_boards variable.

Link: https://wiki.postmarketos.org/wiki/ASUS_Chromebook_C201_(google-veyron-speedy)#U-Boot