ci: add support for integration tests
Replace use of the deprecated "only" keyword with the much more flexible
"rules". Introduce support for spawning a dynamic child pipeline to run
tests against packages that were built in CI.

The child pipeline initially allows for packages to define what tests
and for which devices to run. Initially, this supports the following
kinds of tests:

1. Install tests
"pmbootstrap install" is run with the newly built package in the local
repo.

2. Upgrade tests
Where a chroot for the device is created, the new package made available
via a local repo, and then "apk upgrade -a" is run inside the chroot.
This allows for validating the .post-upgrade script in the context of a
real rootfs.

This is currently limited to packages that are automatically pulled in
as part of a default installation. This is because handling non-default
packages requires either spawning jobs per-package, or assuming that
multiple packages updated in the same MR can all be installed at once
(often not the case). In the end this can't be handled properly without
a dependency graph solver, or by imposing some strong restrictions.
Hence it is deemed out of scope for this initial implementation.

Signed-off-by: Caleb Connolly <caleb@connolly.tech>