modem/rmtfs: add rmtfs-copy-files
This is a helper service which copies the fsc, fsg, modemst1, and
modemst2 partitions during boot so that they can safely be exposed with
R/W permissions to the modem.

If the target directory already exists then copying is skipped.

Intended to be used with the "rmtfs_files_path" setting in
/etc/conf.d/rmtfs.

Signed-off-by: Caleb Connolly <caleb@connolly.tech>