soc-qcom-sdm845: make firmware files RW
Enable the rmtfs-copy-files service and ship a custom /etc/conf.d/rmtfs
which allows the modem to write back its changes to the EFS. Note that
this won't result in changes to the original EFS partition(s) on
internal storage, only to the postmarketOS-specific copy.

It's possible that writing back changes could cause issues with Android,
and until further research/testing this is something we should avoid.

Signed-off-by: Caleb Connolly <caleb@connolly.tech>