main/devicepkg-dev: add support for additional fragments from kernel packages (MR 4675)
Closes #2493

My intention is to implement it like we do for "deviceinfo" file in
device packages. I want maintainrs to be able to just place a
`*.config` file, specify it in $srcdir in APKBUILD and let
kernel_prepare.sh handle everything else.

It is possible to have any amount of files with any names matching
`*.config` pattern.

[ci:skip-build] already built successfully in CI