pmos-ui-gnome: add some useful pkgs to pmb_recmds, drop riscv64
gnome-calendar is unavailable on riscv64, so support for that arch is
dropped until someone is able to resolve that. IMHO this is an ok
trade-off since I don't think many folks are running gnome desktop on
riscv64 (yet?) and gnome-calendar is probably useful for more folks to
have by default.